Macro for Constants (#63)
Remaining an issue: `::` replacement in C++ clashing with ALG type specifier.
