+++ test/MIMO_real.sim
@@ -1,9 +1,9 @@
 # MIMO_real.sim
 # Basic mmWave MIMO model
 # Author: Wuqiong Zhao
-# Date: 2023-04-06
+# Date: 2023-05-04
 
-version: 0.2.0 # the targeted mmCEsim version
+version: 0.2.1 # the targeted mmCEsim version
 meta: # document meta data
   title: mmWave MIMO model
   description:
@@ -12,7 +12,7 @@ meta: # document meta data
   email: contact@mmcesim.org
   website: https://mmcesim.org
   license: MIT
-  date: "2023-04-06"
+  date: "2023-05-04"
   comments: This is an uplink channel.
 physics:
   frequency: narrow # assume narrow band
@@ -52,6 +52,7 @@ sounding:
     channel: "H_cascaded" # the cascaded channel (actually the same as 'H' for simple MIMO)
 preamble: |
   COMMENT Here starts the preamble.
+  COMMENT CHS.PATHS_NUM: `CHS.PATHS_NUM`
   LOG [INFO] A user-defined LOG message.
 estimation: |
   VNt::m = NEW `DICTIONARY.T`
